Expose the ephemeral cluster delete option in the templates (#30)
* Expose the ephemeral cluster delete option in the templates

The oshinko app launcher supports deleting ephemeral clusters,
but the option was not exposed in the templates. This option
has been added and default to "yes". If a cluster is created
on-demand for an app, the cluster will be deleted when the app
completes unless the value is changed.

* Set the ephemeral parameter to be required

This prevents the user from erasing the default "yes" and
leaving the box blank. Changes to the start.sh script will
interpret an unset env var as "yes", so make the user
be explicit.

* Change the affirmative value to "true" from "yes"